Fixes double to float and timer playing by default.
These changes are made to get 3MotionExamples working again.

+++ TDefine.h
@@ -7,7 +7,7 @@
 #define THREEMOTION_TDEFINE_H
 
 namespace ThreeMotion {
-    typedef float milliseconds;
+    typedef double milliseconds;
 }
 
 #endif //THREEMOTION_TDEFINE_H
